pyproject.toml
[tool.poetry]
name = "pdf-chapter-comparator"
version = "0.1.0"
description = "Compare chapters between PDF documents using Azure OpenAI embeddings"
authors = ["Matthias Neumayer <hello@fragdaspdf.de>"]
readme = "README.md"
packages = [{include = "pdf_chapter_comparator"}]

[tool.poetry.dependencies]
python = "^3.11"
pdfplumber = "^0.10.3"
nltk = "^3.8.1"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
tqdm = "^4.66.1"
numpy = ">=1.25.0"
sentence-transformers = "^2.5.1"
torch = ">=2.0.0"
einops = "^0.8.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=42", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pdf-chapter-comparator = "pdf_chapter_comparator.main:main"
